Keep order of static elements in a flow
Before this change, static elements are once removed after the layout, and added back to the end of the flow. Order of elements can vary per page with this implementation. For example, if several elements are specified `-epubx-flow-options: static exclusive last` (used to emulate running headers), the order of those elements are rotated per page, which is not an expected behavior.

The code is changed so that the order of elements in a flow is preserved per page.