Implement UI states selectors (`:checked`, `:enabled` and `:disabled`)
This implementation simply uses attributes on the source DOM node to judge the UI state. Specifically, `:checked` matches if the element has `checked` or `selected` attribute and its value is true, and `:enabled`/`:disabled` matches if the element has `disabled` attribute and its value is false/true. The selectors do not match an element without UI states (such as a `p` element), even if it has `disabled` or `selected` content attribute, for such an element does not have `disabled` or `selected` DOM property.