Fix parse error when '+' follows a type selector and a pseudo-class selector
When a type selector is followed by a pseudo-class selector, the parser cannot distinguish it from a property declaration and tries to judge by looking at the next token. If the next token is, for example, '~', the parser knows it is not a property declaration since a tilde cannot be in a property declaration. However, if the next token is '+', it is allowed in a property declaration *only if* it is followed by a number. The old implementation checked this, and raised an error if the '+' was followed by something other than a number.
In this commit, the code is changed so that if the '+' is followed by something other than a number and if the parser is trying to distinguish a property declaration and a selector, the parser goes back to the start of the selector and parses the input again as a selector.