Fix a page at the viewport origin during the layout
Since a page was attached under a container element with `display: flex` (used for zooming), the origin of the page elements took a random value depending on which page the viewer was displaying. This random offset was affecting float position calculation.
A page element is now fixed at the viewport origin during the layout. Once the layout is done, the page element is re-attached to the container element used for zooming.