Basic support of left/right/recto/verso values for (page-)break-before/after property
start-side of a flow (page side on which the next portion of the flow will be restarted) is resolved from break-before/after values during layout.
The resolved start-side value is used to judge if the flow has contents available for the next page. If the start-side does not match the side of the next page, the flow is regarded as not having available contents.
Note that (page-)break-before value specified on a flow element is not effective yet since in that case the start-side cannot be resolved during layout. The start-side need to be resolved during style calculation before page master selection.