Support (page-)break-before property specified on a flow element
A `break-before` value for a flow element is resolved during style calculation and attached to a FlowChunk object.
If the first-child box of the flow element has a `break-before` value, it is propagated to the FlowChunk. This propagation is repeated recursively at the block edge.
`::before` and `::after` pseudoelements are not considered in this commit. They will be taken into account in the next commit.