Fix handling of float and clear when the element is specified `position: relative`
`NodeContext` had `floatContainer` property, which mixed two different concepts:
- The node establishes a new block formatting context.
- The node establishes a containing block for descendant absolutely positioned elements.

In this fix, `floatContainer` is removed and properties corresponding to those two concepts (`establishesBFC` and `containingBlockForAbsolute) are added.
When a float is inside a block formatting context (except one from the root), float and clear can be controlled by the browser and we don't need to handle these.
When a float is inside a descendant of an element which establishes a containing block for absolutely positioned elements, the position of the float calculated in `layoutFloat` method are converted to one relative to the containing block.