Fix internal hyperlinks in output PDF
Internal hyperlinks did not work in output PDF since IDs on elements were dropped during layout.
To address this issue, such an element is attached a transformed ID, which is given by combining the URL of the document HTML and the original ID and then encoding it by `encodeURIComponent`. This transformation is introduced to keep uniqueness of IDs when there are multple input HTML files (e.g. EPUB).
`href` attribute of an `a` element is also transformed in the same way so that internal hyperlinks in output PDF work correctly.
Logic for page navigation of JS Viewer is also modified.