Fix incorrect page reference in some cases
If an element with an ID does not fit in a page, the view node is removed and deferred to the next page.
However, the ID and the element are registered to an `adapt.vtree.Page` object during creation of the view node and the regitration has not been removed even if the element does not fit in the page.
This has been causing incorrect page references.
In this commit, the registered IDs are checked after layout of the page.
If the page container does not contain an element (elements) corresponding to a registered ID, the registration of the ID is removed.