Prevent reference target from going back to the previous page by reference resolution
If a reference target is allowed to go back to a page earlier than one on which it was placed previously, reference resolution can lead to an infinite loop.
To avoid infinite loops, a reference target is allowed to go to a page later than one on which it was placed previously, but it is not allowed to go back to an earlier page.
If the target can fit in an earlier page, a page break is inserted just before the target and the target is laid out on the page on which it was placed previously.

To represent the influence reference resolution has on layout, `adapt.layout.LayoutConstraint` interface is introduced.
`allowLayout` method of `LayoutConstraint` receives a `NodeContext` and returns whether it can be laid out at the current position.
In this case, a `NodeContext` is not allowed to be laid out at the current position if the page index of the current position is earlier than the index of the page on which the target was placed previously.