Introduced `adapt.layout.BlockFormattingContext` for traditional block layout
A null value of `NodeContext#formattingContext` used to represent that the node is inside a block formatting context.
Now `NodeContext#formattingContext` always has a non-null value, and it is a `BlockFormattingContext` instance when the node is inside a block formatting context.
A new instance of `BlockFormattingContext` is created when the node establishes a new block formatting context.