Fix handling of deferred floats preceded by already laid out floats
If an already laid out page float has a preceding (i.e. has a smaller order) page float which is deferred to the next container, it should be forbidden on the current container, since otherwise the order of them gets reversed.
This decision used to be made at the time when the float is laid out, but it was sometimes causing unnecessarily deferred floats.
For example, when a column page float was deferred from the first column to the second column, a page page float whose anchor appeared in the first column got deferred to the next page, even if the column page float could be laid out in the second column together with the page page float.
With this commit, the decision is not made at the layout time but at the point where the corresponding page float layout context is ‘finish’ed.
An `isAllowedToPrecede` method is also added to PageFloat, which enables skipping the above decision.
The method is overridden by Footnote since it is preferable not to defer footnotes even if some preceding page floats (which are not footnotes) are deferred.