Fix bug on epubcfi failing to navigate to beginning of a spine item
The fragment parameter `f=epubcfi(…)` of vivliostyle-viewer URL did not work when trying to navigate to beginning of an EPUB spine item (e.g., beginning of a chapter of a publication).

For example, this vivliostyle-viewer URL using the `f=epubcfi()`
http://vivliostyle.github.io/vivliostyle.js/viewer/vivliostyle-viewer.html#b=/vivliostyle_doc/ja/events/html5jpub2017/&spread=false&f=epubcfi(/4/6!)
is referring the beginning of the 3rd item ("Webと出版が融合？"),
but due to this bug the navigation failed and the first page of the publication opened.