Fix error "Failed to fetch a source document" with web publications on Microsoft Edge
The problem was: The document URLs from TOC items are obtained by `anchorElem.href` property.
This href value is absolute URL that is generated from the "href" attribute value and the base URL.
On Chrome, Safari, and Firefox, this base URL is the URL of the document containing the TOC,
and works as expected. However, on Microsoft Edge, the vivliostyle-viewer URL is used as the base URL
and the `anchorElem.href` value becomes wrong URL, and causes "Failed to fetch a source document"
error.