refactor: Revert 'refactor: use "for-of" instead of "forEach"'
The change in 'use "for-of" instead of "forEach"' 92c7349 of PR #658 had problem.

`for (const element of array) { ... }` is not equivalent to `list.forEach((item) => { ... });` when the list may have empty elements!
"forEach" skips empty elements but "for-of" does not skip and the `element` value becomes `undefined`,
so when accessing a property of `element`,  "TypeError: Cannot read property 'xxx' of undefined" occurs.