chore: remove "npm-next" link in README
The npm tag 'next' was used for pre-release versions, but currently there are no new pre-release versions and the latest release version is newer than the last pre-release version.

Note: for testing the latest development version, use the [Canary Release](https://vivliostyle.github.io/#canary-release-equivalent-to-master).

We will use the 'next' tag again when preparing next major version, but for now it should be removed to avoid confusion ('next' is old).

I did also remove 'next' tag in the npm packages:
```
$ npm dist-tag rm @vivliostyle/viewer next
$ npm dist-tag rm @vivliostyle/core next
$ npm dist-tag rm @vivliostyle/react next
```