feat: Support printing mixed page sizes
- resolves #751

This change enables mixed page sizes in output PDF together with the following change in vivliostyle-cli:

- https://github.com/vivliostyle/vivliostyle-cli/pull/278

When multiple page sizes are specified, as in the example in Issue 751,
the largest page size both vertically and horizontally are used in PDF output
from Vivliostyle.js.
Then the PDF is post-processed in the Vivliostyle CLI to set the multiple page sizes as specified.