fix: CSS parser error with property:calc(…) misinterpreted as selector with pseudo class
- fix #1020

Before this fix, when starting parsing a CSS rule body, the token sequence `<ident>` +  `:` + `<ident>` was interpreted as property assignment initially and reinterpreted as selector with a pseudo class when unexpected property value syntax was found.
The problem was that the property value syntax check was incomplete, especially with calc() expression , and that caused misinterpretation of property:calc(…) as selector with pseudo class and it resulted in the CSS parser error.

In this fix, `inStyleDeclaration: boolean` property is added in the `class Parser` and `inStyleDeclaration` is managed so that it is true when property asignment is expected and selector is not.