followup fix for #1079: add pixelRatioLimit
Pixel ratio emulation on PDF output (PR #1079) does not work with
non-Chromium browsers and has printing problem with Microsoft Edge
(negative page margins used in this emulation causes pages disappeared).

So limit this feature for Chromium browsers except Microsoft Edge.
Sets the max pixelRatio value pixelRatioLimit to 16 for Chromium browsers except Microsoft Edge,
and 0 (disabled) for non-Chromium browsers and Microsoft Edge.
(The defualt value of pixelRatio is 8, if not disabled)