fix: very thin border on printing with Microsoft Edge
When the pixel ratio emulation for thin border width was implemented (PR #1079), it did not work well with Microsoft Edge on printing, so this feature was disabled for Microsoft Edge (PR #1085).

Now I tested it again with the latest Microsoft Edge (109.0.1518.61) and it works fine.

This feature is now enabled on all Chromium browsers including Microsoft Edge.