fix list style in the user agent default style sheet
Before this fix, the list margin style was defined as follows:

```css
html|ul,
html|ol,
html|dl,
html|dir,
html|menu {
  margin-block-start: 1em;
  margin-block-end: 1em;
  margin-inline-start: 0em;
  margin-inline-end: 0em;
}
...

html|ol,
html|ul,
html|dir,
html|menu {
  margin: 0px;
  margin-inline-start: 40px;
}
```

Because of the bug #1126, the `margin-inline-start: 40px` was ignored, and browser's default `padding-inline-start: 40px` was effective.

This causes problem after this bug was fixed, because both `margin-inline-start: 40px` and the browser's default `padding-inline-start: 40px` become effective,
so the `margin-inline-start: 40px` should be changed to `padding-inline-start: 40px` which is consistent with the browser's default.