fix: tweak TOC detection for VFM v2 change
This resolves the problem on VFM v2 change:

- https://github.com/vivliostyle/vfm/pull/157

With the following Markdown,

```md
 # Table of Contents {.toc}

- [One](one.html)
- [Two](two.html)
```

VFM v1 generates the following HTML:

```html
    <section class="level1 toc" id="table-of-contents">
      <h1 class="toc">Table of Contents</h1>
      <ul>
        <li><a href="one.html">One</a></li>
        <li><a href="two.html">Two</a></li>
      </ul>
    </section>
```

Vivliostyle.js recognizes the section as TOC because "toc" class is found in the section element.

This is changed in VFM v2. It generates the following HTML:

```html
    <section class="level1" aria-labelledby="table-of-contents">
      <h1 class="toc" id="table-of-contents">Table of Contents</h1>
      <ul>
        <li><a href="one.html">One</a></li>
        <li><a href="two.html">Two</a></li>
      </ul>
    </section>
```

The "toc" class is not found in the section element, so the TOC detection has to be tweaked using the selector,
`section:has(>:first-child:is(h1,h2,h3,h4,h5,h6):is(.toc,#toc))`.

Note: the stylesheet for TOC that was a part of UserAgentBaseCSS is now a separate stylesheet UserAgentTocCSS.
This change is necessary because uaStylesheetBaseFetcher cannot handle the `:has()` selector.