feat(viewer): Add UI Language in settings panel
- Add UI Language in settings panel
- New configuration flag "L": disable UI Language change in the setting panel

When Viewer is open, the UI language is determined with the following priority order:

1. Check URL parameter (in query or fragment): "?lng=xx", "&lng=xx" or "#lng=xx"
2. Check LocalStorage with "i18nextLng" key
3. browser navigator.language value

When the UI language is changed in the settings panel, it is saved to the LocalStorage with "i18nextLng" key.

Currently, only two languages, "en" and "ja", are available. English (en) is used as the fallback language.