fix: `-adapt-template: footnote` should work for "user-agent.xml#footnote"
So far `-adapt-template: url(user-agent.xml#footnote)` is used for EPUB footnote
in the UA default style sheet. However, this feature could not be used easily
in user or author style sheets because the "user-agent.xml" is bundled in the
"assets.ts" file.

This fix enables `-adapt-template: footnote` that should work same as
`-adapt-template: url(user-agent.xml#footnote)` in the UA default style sheet.