fix: Fix auto spread view behavior when page size changes
In Vivliostyle Viewer, with auto spread view mode, when the page size changes in the middle of the document from portrait to landscape, the spread behavior should be changed to single page view. However, the spread behavior was not changed and the pages were displayed side by side, which made it difficult to read.