fix(viewer): Do not remove console output in vivliostyle-viewer.js
So far, `packages/viewer/rollup.config.js` has removed `console.*` output in release build.
This is a problem because especially in Vivliostyle CLI, the browser console output is used for error detection.

This problem has become apparent since v2.30.6 (Vivliostyle CLI v8.16.2).
The reason is that in v2.30.6, the build of vivliostyle.js (core) was changed from microbundle to esbuild.

- https://github.com/vivliostyle/vivliostyle.js/pull/1417

In the JS code generated by microbundle, the original source code `console.error(...msg)` was transformed into `(n=console).error.apply(n,t)`, so when vivliostyle-viewer.js was built with vivliostyle.js (core) bundled, that part was not removed.
However, in the JS code generated by esbuild, that transformation is not done, so when vivliostyle-viewer.js is built with vivliostyle.js (core) bundled, that part is removed.

This commit stops removing `console.*` output in release build.