fix: Fix EPUB footnote (epub:type="footnote") not working (Regression in v2.31.2)
This bug was a side effect of the change to use the fetch API instead of XMLHttpRequest for fetching documents (PR #1461).

The problem was that the `URL` of the parsed document was set to the URL of the Vivliostyle Viewer with fragment `#src=...`, instead of the original document URL, because the `DOMParser.parseFromString()` method sets the `URL` of the parsed document to the URL of the current `window.document`. This caused the `elem.href == elem.baseURI + elem.hash` check to fail, preventing the EPUB footnote processing from working.

- fix #1521